Modified multiple views to make them look better.
Developed a component with the detail of the brand and other similar component with the detail of the motorcycle, then included these componentes in all of the 'show' views.

Included the search bar component in the admin views.

Modified the navbar by adding a nav item that redirects to the shopping cart (at this moment there was no way to get into this module).