Changes in order's logic.
Order logic was presenting a problem since we were substracting a quantity from the motocycle stock before we knew if the user could pay the total amount of the order. Also, there was another problem with adding the same motorcycle to the cart but in different instances, since we were not validating that the sum of the quantities of this same motorcycle was less than the stock of it.